+++ packages/archives/src/professional/components/experience/index.ts
@@ -1,1 +1,2 @@
-export {default as ExperienceSection} from "./experienceSection";
+export { default as ExperienceBadge } from "./experienceBadge";
+export { default as ExperienceSection } from "./experienceSection";