+++ packages/archives/src/professional/index.ts
@@ -1,5 +1,6 @@
 export {
 	CodeEditorConsoleFrame,
+	ExperienceBadge,
 	ExperienceSection,
 	HelloworldInteractiveHero,
 	MatrixInteractiveHero,