+++ packages/curio/src/components/badge/index.ts
@@ -1,0 +1,3 @@
+export { default as Badge } from "./badge";
+
+export type { BadgeProps } from "./badge";