+++ packages/curio/src/components/index.ts
@@ -1,5 +1,6 @@
 export * from "./alert";
 export * from "./backdrop";
+export * from "./badge";
 export * from "./button";
 export * from "./card";
 export * from "./chip";